super-admin-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
